// src/@types/multi-tracks-encoder.d.ts
import { User } from "eris";
import { Chunk } from "./audio-recorder";

export interface IRecordingDetails {
  guild: string;
  channel: string;
}
export interface IMultiTracksEncoder {
  initStreams(recordId: string, details: IRecordingDetails): void;
  registerNewTrackForUser(userTrackNo: number, user: User): void;
  encodeChunk(
    user: User,
    streamNo: number,
    packetNo: number,
    chunk: Chunk
  ): void;
  flush(
    user: User,
    streamNo: number,
    queue: Chunk[],
    ct: number,
    packetNo: number
  ): number;
  closeStreams(): void;
}